rental_etl/requirements.txt
apache-airflow==2.10.5
SQLAlchemy==1.4.54
pymongo==4.11.1
psycopg2-binary==2.9.10
